Project/backend/requirements.txt
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.1
gast==0.3.3
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.2.0
